pyproject.toml
[project]
name = "X_IL"
description = "X-IL: Exploring the Design Space of Imitation Learning Policies"
readme = "README.md"
requires-python = ">=3.10,<3.11"
version = "0.1.0"
dependencies = [
    "hydra-core==1.3.2",
    "numpy",
    "wandb",
    "easydict",
    "transformers",
    "opencv-python",
    "einops",
    "thop",
    "robosuite==1.4.0",
    "bddl",
    "future",
    "matplotlib",
    "cloudpickle",
    "gym",
    "numexpr",
    "h5py",
    "torchsde",
    "torchdiffeq",
    "imageio[ffmpeg]",
    "open3d",
    "ftfy",
    "omegaconf",
    "Pillow",
    "regex",
    "scipy",
    "termcolor",
    "torch==2.5.1",
    "torchvision==0.20.1",
    "tqdm",
]

license = { file = "LICENSE" }
classifiers = ["License :: OSI Approved :: MIT License"]


[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.format]
quote-style = "single"

[tool.basedpyright]
pythonVersion = "3.10"
typeCheckingMode = "basic"

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true
